Move water TEDs to their own file
