+++ src/openrct2/libopenrct2.vcxproj
@@ -568,6 +568,7 @@
     <ClInclude Include="ride\ted\TED.MiniGolf.h" />
     <ClInclude Include="ride\ted\TED.SBend.h" />
     <ClInclude Include="ride\ted\TED.Station.h" />
+    <ClInclude Include="ride\ted\TED.Water.h" />
     <ClInclude Include="ride\ted\TED.ZeroG.h" />
     <ClInclude Include="ride\ted\TrackElementDescriptor.h" />
     <ClInclude Include="ride\ted\TrackElemType.h" />
